change big ints to ints due to axios issues
  * axios can not parse bigints

+++ server/src/main/java/com/taskflow/server/dtos/GroupDTO.java
@@ -10,6 +10,6 @@ public class GroupDTO {
     private int id;
     private String name;
     private LocalDateTime creationDate;
-    private Long position;
+    private int position;
 
 }
