+++ server/src/main/java/com/taskflow/server/entities/Task.java
@@ -31,7 +31,7 @@ public class Task {
     private LocalDateTime creationDate;
 
     @Column(nullable = false)
-    private long position;
+    private int position;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "board_id", nullable = false, insertable = false, updatable = false)
